mod: bump dependencies version fix??: `SmbjProxyFileCallbackSafe::onWrite` return type mismatch fix: pin shared dependency `org.bouncycastle:bcprov-jdk15to18:1.76`

+++ data/storage/smbj/src/main/java/com/wa2c/android/cifsdocumentsprovider/data/storage/smbj/SmbjProxyFileCallbackSafe.kt
@@ -42,7 +42,8 @@ class SmbjProxyFileCallbackSafe(
     override fun onWrite(offset: Long, size: Int, data: ByteArray): Int {
         if (mode != AccessMode.W) { throw ErrnoException("Writing is not permitted", OsConstants.EBADF) }
         return processFileIo(coroutineContext) {
-            file.write(data, offset, 0, size)
+            // FIXME: long value clamped to int
+            file.write(data, offset, 0, size).toInt()
         }
     }
